initial API impl
